requirements.txt
pytest==7.4.3
selenium==4.15.2
webdriver-manager==4.0.1
pytest-html==3.2.0
pytest-xdist==3.2.1
requests==2.31.0
Faker==20.0.3
psycopg2-binary==2.9.9